requirements.txt
# Core
python-pptx>=0.6.23
pandas>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# AI/LLM
google-genai>=1.0.0

# Web/UI
streamlit>=1.30.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utils
loguru>=0.7.0
tenacity>=8.2.0
